requirements.txt
streamlit
python-dotenv
selenium
webdriver-manager

langchain
langchain_core
langchain-community
langchain_google_genai


playwright
sentence-transformers

faiss-cpu
unstructured